File_Organizer/requirements.txt
Flask==2.3.3
google-generativeai==0.3.2
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3